third_party/WebKit/LayoutTests/fast/parser/foster-parent.html

<script src="../../resources/js-test.js"></script>

<div id="div">
  <table id="table">
    <script>
      var table = document.getElementById('table');
      table.parentNode.removeChild(table);
      var docFragment = document.createDocumentFragment();
      docFragment.appendChild(table);
    </script>
    <h1></h1>
  </table>
</div>

<script>
description("Tests the HTML foster parenting algorithm when a DocumentFragment is the chosen foster parent");
var div = document.getElementById('div');

shouldBe("div.children.length", "0");
shouldBe("docFragment.firstElementChild.tagName", "'H1'");
</script>
